Decimal type for currency amounts.

+++ models/fund_transaction.py
@@ -3,6 +3,6 @@ from models.fund import Fund
 
 class FundTransaction(mongoengine.EmbeddedDocument):
     fund = mongoengine.ReferenceField(Fund, required=True)
-    change = mongoengine.FloatField(required=True)
-    assigment = mongoengine.FloatField(required=True)
+    change = mongoengine.DecimalField(required=True)
+    assigment = mongoengine.DecimalField(required=True)
     balance = mongoengine.FloatField(required=True)
